fix(runtime): avoid acceptance callback import cycle
